requirements.txt
requests==2.27.1
beautifulsoup4==4.10.0
bs4==0.0.1
python-dotenv==0.19.2
tweepy==4.5.0
